requirements.txt
setuptools
slack_sdk
aiohttp
python-dotenv
